% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_CaseReserve_calc.R
\name{create_CaseReserve}
\alias{create_CaseReserve}
\title{This Function takes two triangles (Reported and Paid) and calculates the Case
 Reserve triangle by taking the difference reported and paid triangle.}
\usage{
create_CaseReserve(ReportedTriangle, PaidTriangle)
}
\arguments{
\item{ReportedTriangle}{The reported triangle (as a ChainLaddder package triangle)}

\item{PaidTriangle}{The paid triangle (as a ChainLaddder package triangle)}
}
\value{
CaseReserve, triangle (ChainLadder package type)
}
\description{
This Function takes two triangles (Reported and Paid) and calculates the Case
 Reserve triangle by taking the difference reported and paid triangle.
}
