% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_CaseReserves_fromTriangles.R
\name{create_casereserves_fromtriangles}
\alias{create_casereserves_fromtriangles}
\title{This Function takes a nested list of the following structure
[[portfolio name]][[type of names]] which should include any kind of reported
value and possibly paid values (if paid is missing, we would just take
reported == case reserves)}
\usage{
create_casereserves_fromtriangles(nested_list)
}
\arguments{
\item{triangle_list}{Nested list of the following structure
[[portfolio name]][[type of names]] having at least reported triangle values in it.}
}
\value{
triangle_list, list of all triangles which is structured
     by [[portfolio_name]][[type_of_amount]]. Additionally, the Case Reserve
     values are included
}
\description{
This Function takes a nested list of the following structure
[[portfolio name]][[type of names]] which should include any kind of reported
value and possibly paid values (if paid is missing, we would just take
reported == case reserves)
}
