% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateTriangles_from_DataTable.R
\name{create_triangle_fromdata}
\alias{create_triangle_fromdata}
\title{This Function takes a data table (data) and creates from it for each
type of amount and portfolio name triangles, so that one could do triangle
manipulations (as differences between triangles or conversion from cumulative
to incremental triangles)}
\usage{
create_triangle_fromdata(datamat, cumorinc)
}
\arguments{
\item{datamat}{Data frame/table that contains at least following columns:
- Portfolio Name
- Type of Amount
- Development Period
- Origin Period
- Amount}

\item{cumorinc}{Boolean (T/F), which tells us if the data is given as
cumulative values (cumorinc = T) or as incremental values (cumorinc = F)}
}
\value{
triangle_list, list of all triangles which is structured
     by [[portfolio_name]][[type_of_amount]]
}
\description{
This Function takes a data table (data) and creates from it for each
type of amount and portfolio name triangles, so that one could do triangle
manipulations (as differences between triangles or conversion from cumulative
to incremental triangles)
}
