% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Excel_functions.R
\name{fileinp.filereadin}
\alias{fileinp.filereadin}
\title{Function that loads all the information coming from the Shiny-widget
fileInput() into a list and prepares/processes all the respective elements.
Using this function, one could also upload multiple files using the named widget
and gets a nested list of all the information.}
\usage{
fileinp.filereadin(fileinp, shtnms, range.selection, mltple)
}
\arguments{
\item{fileinp}{Uploaded file information using fileinput() like name, temp. path etc.}

\item{shtnms}{Excel-Sheet names of the uploaded file/-s}

\item{range.selection}{Range which should be extracted from Excel. Default is selected NULL, i.e. no selection.}

\item{mltple}{Are multiple Files uploaded? If yes, TRUE is selected}
}
\value{
List with the following information:
- df              = Uploaded data frame
- extension       = Ending of the data file (e.g. .xlsx)
- filepath        = Temporary file path
- fliename        = Complete file name
}
\description{
Function that loads all the information coming from the Shiny-widget
fileInput() into a list and prepares/processes all the respective elements.
Using this function, one could also upload multiple files using the named widget
and gets a nested list of all the information.
}
