% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates_bysum.R
\name{remove_duplicates_bysum}
\alias{remove_duplicates_bysum}
\title{Reads in a dataframe and checks for duplicates with respective to SPIREs uniqueness key.
This uniqueness key consists of the following columns (separated by "_")
- Portfolio Name
- Origin Period
- Development Period
- Type of Amount
In case some duplicated rows are detected, the function sums up the duplicated values and
return a version of the initial dataframe with only one representative of the duplicated rows.
This representative contains as amount the sum of the amounts.}
\usage{
remove_duplicates_bysum(df_to_manipulate)
}
\arguments{
\item{df_to_manipulate}{Dataframe in which we are looking for the duplicates.}
}
\value{
result_df Dataframe with no duplicated rows.
}
\description{
Reads in a dataframe and checks for duplicates with respective to SPIREs uniqueness key.
This uniqueness key consists of the following columns (separated by "_")
- Portfolio Name
- Origin Period
- Development Period
- Type of Amount
In case some duplicated rows are detected, the function sums up the duplicated values and
return a version of the initial dataframe with only one representative of the duplicated rows.
This representative contains as amount the sum of the amounts.
}
\examples{
result_df <- remove_duplicates_bysum(df_to_manipulate = result_df)
}
