% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform_to_Template.R
\name{transform_to_template}
\alias{transform_to_template}
\title{Function that loads all the information coming from the Shiny-widget
fileInput() into a list and prepares/processes all the respective elements.
Using this function, one could also upload multiple files using the named widget
and gets a nested list of all the information.}
\usage{
transform_to_template(dattab)
}
\arguments{
\item{dattab}{Data table that is uploaded from the user and needs to be renamed
to the corresponding SPIRE template column names. Further the missing columns
that are needed are added.}
}
\value{
List with the following information:
- df              = Uploaded data frame
- extension       = Ending of the data file (e.g. .xlsx)
- filepath        = Temporary file path
- fliename        = Complete file name
}
\description{
Function that loads all the information coming from the Shiny-widget
fileInput() into a list and prepares/processes all the respective elements.
Using this function, one could also upload multiple files using the named widget
and gets a nested list of all the information.
}
